Bug fixes (#2)
* :bug:

It turns out that the solution to the square well potential is giving a
lot of numerical problems.

* :bug:

Fixed the mean spherical approximation closure relation.

* :sparkles: :art:

Formatting, and added a new potential for testing.

* :bug:

Seems like neither potential is working, but it might be the potential
itself. Some more tests are pending, but I really don't know what the
problem is.

* :white_check_mark:

Tests are now passing. They pass when the cutoff radius is small, and
density grid is large enough.

* :white_check_mark:

Reduce the number of grid points in the distance vector.

* :bulb:

* :white_check_mark:

Started testing all the implemented closures, specifically for the
square well potential.

* :recycle: :white_check_mark:

Refactored the closure relations definitions to make some of them
mutable, mostly the ones that have free parameters. Updated the tests to
have most coverage of these closure relations.

* :white_check_mark:

This should include all tests of the current source code.
Bumping version to v1.2

* :art: :coffin:

Formatting using JuliaFormatter. Also, removed the definition of the
ConsConstantClosure that is no longer implemented.